requirements.txt
asgiref==3.8.1
Django==4.2.13
django-ratelimit==4.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
packaging==24.1
PyMySQL==1.1.1
python-dotenv==1.0.1
ratelimit==2.2.1
sqlparse==0.5.0
typing_extensions==4.12.2
